requirements.txt
openai>=1.41.0
pytest-mock
